Finish caesar cipher

+++ basics/alpha.h
@@ -2,4 +2,4 @@
 #include <stdlib.h>
 
 char ALPHABET[] = { 'a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z' };
-size_t LENGTH = sizeof(ALPHABET)/sizeof(char);
+int LENGTH = sizeof(ALPHABET)/sizeof(char);
